packages/bot_tele/requirements.txt
httpx==0.27.0
python-telegram-bot==21.4
chatgpt-md-converter==0.1.1
python-dotenv==1.0.1
